update orderDetail resquest

+++ src/main/java/com/project/shopapp/dtos/OrderDetailDTO.java
@@ -25,7 +25,7 @@ public class OrderDetailDTO {
 
     @Min(value = 1, message = "total_money must be >= 0")
     @JsonProperty("total_money")
-    private int totalMoney;
+    private Float totalMoney;
 
     private String color;
 }
